+++ src/main/java/zenpad/misc/AboutDialog.java
@@ -1,4 +1,4 @@
-package zenpad.ui;
+package zenpad.misc;
 
 import javax.swing.ImageIcon;
